add sql tchart tests

+++ client/setup-icons.js
@@ -3,13 +3,18 @@ const fs = require('fs');
 const path = require('path');
 const icons = require('@mdi/js');
 
-exports.saveMdiIcons = () => {
+const saveMdiIcons = () => {
+  const iconEntries = Object.entries(icons);
   const iconsDestinationFolder = path.join(__dirname, '/static/icons');
   if(fs.existsSync(iconsDestinationFolder)) {
+    const contents = fs.readdirSync(iconsDestinationFolder);
+    if(contents.length >= iconEntries.length) {
+      console.log(`Already saved ${contents.length} icons`);
+      return;
+    }
     fs.rm(iconsDestinationFolder, { recursive: true }, console.log);
   }
   fs.mkdirSync(iconsDestinationFolder, { recursive: true });
-  const iconEntries = Object.entries(icons);
   if(!iconEntries.length) {
     console.error('No icons found. Did you run npm i ?!');
     process.exit(1);
@@ -24,5 +29,27 @@ exports.saveMdiIcons = () => {
     </svg>`;
     fs.writeFileSync(path.join(iconsDestinationFolder, `${name.slice(3)}.svg`), iconSvg, { encoding: 'utf8' });
   });
+  console.log(` Saved ${iconEntries.length} icons`);
   fs.writeFileSync(path.join(iconsDestinationFolder, "_meta.json"), JSON.stringify(iconNames, null, 2), { encoding: 'utf8' });
 }
+
+setTimeout(saveMdiIcons, 1000);
+
+class SaveMdiIcons {
+  apply(compiler) {
+    compiler.hooks.afterEmit.tap(
+      'Save Mdi icons plugin afterEmit',
+      _stats => {
+        setTimeout(saveMdiIcons, 1000);
+      }
+    );
+    compiler.hooks.done.tap(
+      'Save Mdi icons plugin done',
+      _stats => {
+        setTimeout(saveMdiIcons, 1000);
+      }
+    );
+  }
+}
+
+module.exports = { SaveMdiIcons };
